+++ spec/integration_spec.rb
@@ -7,8 +7,8 @@ describe "CLI" do
     File.expand_path(File.join(File.dirname(__FILE__),'..','bin','bundle-audit'))
   end
 
-  context "when auditing a vulnerable bundle" do
-    let(:bundle)    { 'vuln' }
+  context "when auditing a bundle with unpatched gems" do
+    let(:bundle)    { 'unpatched_gems' }
     let(:directory) { File.join('spec','bundle',bundle) }
 
     subject do
@@ -23,7 +23,7 @@ describe "CLI" do
       subject.should include(%{
 Name: actionpack
 Version: 3.2.10
-CVE: 2013-0156
+Advisory: CVE-2013-0156
 Criticality: High
 URL: http://osvdb.org/show/osvdb/89026
 Title: Ruby on Rails params_parser.rb Action Pack Type Casting Parameter Parsing Remote Code Execution
@@ -31,7 +31,7 @@ Solution: upgrade to ~> 2.3.15, ~> 3.0.19, ~> 3.1.10, >= 3.2.11
 
 Name: activerecord
 Version: 3.2.10
-CVE: 2013-0276
+Advisory: CVE-2013-0276
 Criticality: Medium
 URL: http://direct.osvdb.org/show/osvdb/90072
 Title: Ruby on Rails Active Record attr_protected Method Bypass
@@ -39,11 +39,27 @@ Solution: upgrade to ~> 2.3.17, ~> 3.1.11, >= 3.2.12
 
 Name: activerecord
 Version: 3.2.10
-CVE: 2013-0155
+Advisory: CVE-2013-0155
 Criticality: High
 URL: http://osvdb.org/show/osvdb/89025
 Title: Ruby on Rails Active Record JSON Parameter Parsing Query Bypass
 Solution: upgrade to ~> 2.3.16, ~> 3.0.19, ~> 3.1.10, >= 3.2.11
+      }.strip)
+    end
+  end
+
+  context "when auditing a bundle with insecure sources" do
+    let(:bundle)    { 'insecure_sources' }
+    let(:directory) { File.join('spec','bundle',bundle) }
+
+    subject do
+      Dir.chdir(directory) { sh(command, :fail => true) }
+    end
+
+    it "should print warnings about insecure sources" do
+      subject.should include(%{
+Insecure Source URI found: git://github.com/rails/jquery-rails.git
+Insecure Source URI found: http://rubygems.org/
       }.strip)
     end
   end